chore: separate FlowState class from Coordinator
